# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample swagger for Fiber",
        "title": "Finance Forge Server",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "fiber@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "/",
    "paths": {
...
        "/api/token/allowance": {
            "get": {
                "description": "Returns the remaining number of tokens that the spender is allowed to spend on behalf of the owner",
                "tags": [
                    "token"
                ],
                "summary": "Retrieve remaining token allowance",
                "parameters": [
                    {
                        "type": "string",
                        "description": "owner address",
                        "name": "owner",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "spender address",
                        "name": "spender",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful retrieval",
                        "schema": {
                            "$ref": "#/definitions/handlers.AllowanceResponse"
                        }
                    },
                    "404": {
                        "description": "Allowance not found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/token/approvals": {
            "get": {
                "description": "Returns all the approvals that have happened on the blockchain and captured by the server",
                "tags": [
                    "token"
                ],
                "summary": "Retrieve all approvals",
                "responses": {
                    "200": {
                        "description": "Successful retrieval",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Approval"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/token/balance/{account}": {
            "get": {
                "description": "Returns the value of tokens owned by the specified account",
                "tags": [
                    "token"
                ],
                "summary": "Retrieve account balance",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Account address",
                        "name": "account",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful retrieval",
                        "schema": {
                            "$ref": "#/definitions/handlers.BalanceResponse"
                        }
                    },
                    "404": {
                        "description": "Account not found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/token/transfers": {
            "get": {
                "description": "Returns all the transfers that have happened on the blockchain and captured by the server",
                "tags": [
                    "token"
                ],
                "summary": "Retrieve all transfers",
                "responses": {
                    "200": {
                        "description": "Successful retrieval",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Transfer"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handlers.AllowanceResponse": {
            "type": "object",
            "properties": {
                "allowance": {
                    "type": "string"
                }
            }
        },
        "handlers.BalanceResponse": {
            "type": "object",
            "properties": {
                "balance": {
                    "type": "string"
                }
            }
        },
...
        "handlers.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "models.Approval": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deleted_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "owner": {
                    "type": "string"
                },
                "spender": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        },
...
        "models.Transfer": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deleted_at": {
                    "type": "string"
                },
                "from": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "to": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        },
...
    }
}